+++ AddIns/WebLogAddin/MetaWebLogApi/classes/CustomField.cs
@@ -1,8 +1,11 @@
-﻿namespace WebLogAddin.MetaWebLogApi
+﻿using System.Diagnostics;
+
+namespace WebLogAddin.MetaWebLogApi
 {
     /// <summary>
     /// Custom field info attached to a blog item.
     /// </summary>
+    [DebuggerDisplay("{Key},{Value}")]
     public class CustomField
     {
         public string ID { get; set; }